pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "nullsec-hak5-toolkit"
version = "1.0.0"
description = "Universal command-line toolkit for the Hak5 device ecosystem"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "NullSec", email = "badxantics@gmail.com"}
]
keywords = ["hak5", "pineapple", "bashbunny", "rubberducky", "pentest", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "requests>=2.28",
    "pyserial>=3.5",
    "pyusb>=1.2",
    "paramiko>=3.0",
    "netifaces>=0.11",
]

[project.scripts]
hak5 = "core.cli:main"

[project.urls]
Homepage = "https://github.com/bad-antics/nullsec-hak5-toolkit"
Repository = "https://github.com/bad-antics/nullsec-hak5-toolkit"
Issues = "https://github.com/bad-antics/nullsec-hak5-toolkit/issues"
